\name{aqm.plot}
\alias{aqm.plot}
\alias{aqm.plot-methods}
\alias{aqm.plot,aqmTrellis-method}
\alias{aqm.plot,aqmPlotfun-method}

\title{ Make plots from aqmTrellis and aqmPlotfun objects. }
\description{ Make plots from aqmTrellis and aqmPlotfun objects. }

\usage{ aqm.plot(obj) }

\arguments{
  \item{obj}{ an object of class \code{aqmTrellis} or \code{aqmPlotfun}. }
 }
 
\details{ See the aqm Vignette.  }

\author{
  Audrey Kauffmann 
}

