\name{outlierDetection}
\docType{class}

\alias{class:outlierDetection}
\alias{outlierDetection}
\alias{outlierDetection-class}

\title{Represents the results from applying an outlier detection
  criterion to the arrays.}

\description{
The class is described in the vignette \emph{Advanced topics: Customizing arrayQualityMetrics
  reports and programmatic processing of the output}.
}

\seealso{
\code{\link{outliers}}
}

\author{Audrey Kauffmann, Wolfgang Huber}


