\name{aqm.plot}
\alias{aqm.plot}
\alias{aqm.plot-methods}
\alias{aqm.plot,aqmTrellis-method}
\alias{aqm.plot,aqmobj.box-method}
\alias{aqm.plot,aqmobj.dens-method}
\alias{aqm.plot,aqmobj.msd-method}
\alias{aqm.plot,aqmobj.nuse-method}
\alias{aqm.plot,aqmobj.pmmm-method}
\alias{aqm.plot,aqmobj.qcs-method}
\alias{aqm.plot,aqmobj.rle-method}
\alias{aqm.plot,aqmobj.rnadeg-method}

\title{ Performs plots from aqm objects. }

\description{ \code{aqm.plot} performs plots. }

\usage{ aqm.plot(obj) }

\arguments{
  \item{obj}{ an object of class \code{aqmobj} }
 }
 
\details{ See the aqm.prepdata help or the aqm Vignette 
  for example of this object.  }

\value{ A plot in the x11 device. } 

\author{
  Audrey Kauffmann 
  Maintainer: <kauffmann@bergonie.org>  
}

