\name{aqmobj.prepaffy-class}
\docType{class}
\alias{aqmobj.prepaffy-class}
\alias{aqmobj.prepaffy}
\alias{class.aqmobj.prepaffy}

\title{ Class to contain data generated from aqm.prepaffy. }

\description{ Container for the output of \code{\link{aqm.prepaffy}} and
  for the input of the \code{\link{aqm.rle}} and \code{\link{aqm.nuse}}
  functions. }

\section{Slots}{
  \describe{
    \item{\code{dataPLM}:}{ A \code{\link[affyPLM:PLMset-class]{PLMset}}. }
    }
  }
  
\details{
See \code{\link{aqm.prepaffy}} and the \code{aqm} Vignette.
}

\author{ Audrey Kauffmann <kauffmann@bergonie.org> }

\keyword{classes}
