\name{aqmReportModule}
\docType{class}

\alias{class:aqmReportModule}
\alias{aqmReportModule}
\alias{aqmReportModule-class}

\title{Class to contain all the information to render a quality report module.}

\description{
Please see 
the vignette \emph{Advanced topics: Customizing arrayQualityMetrics
reports and programmatic processing of the output}.}

\section{Creating Objects}{
See the manual page of \code{\link[aqm.boxplot]{module functions}}.
}

\author{Audrey Kauffmann, Wolfgang Huber}


