\name{aqmReportModule}
\docType{class}

\alias{class:aqmReportModule}
\alias{aqmReportModule}
\alias{aqmReportModule-class}

\title{Class to contain all the information to render a quality report module.}

\description{
Please see 
the vignette \emph{Advanced topics: Customizing arrayQualityMetrics
reports and programmatic processing of the output}.}

\section{Creating Objects}{
Please see the manual page of the module generations functions, e.g. \code{\link{aqm.boxplot}}.
}

\author{Audrey Kauffmann, Wolfgang Huber}


