\name{outlierDetection}
\docType{class}

\alias{class:outlierDetection}
\alias{outlierDetection}
\alias{outlierDetection-class}

\title{Represents the results from an applying an outlier detection
  criterion to the arrays.}

\description{
Please see the vignette \emph{Advanced topics: Customizing arrayQualityMetrics
  reports and programmatic processing of the output}.
}

\author{Audrey Kauffmann, Wolfgang Huber}


